+++ src/pages/actions/register/topup/DeleteTopupConfirmation.tsx
@@ -3,13 +3,13 @@ import styled from "styled-components";
 import { useTranslation } from "react-i18next";
 import { useDispatch, useSelector } from "react-redux";
 
-import { useBackd } from "../../../app/hooks/use-backd";
-import { AppDispatch } from "../../../app/store";
-import Popup from "../../../components/Popup";
-import { Position, Pool, TransactionInfo, fromPlainPosition } from "../../../lib/types";
-import { removePosition } from "../../../state/positionsSlice";
-import { selectTransactions } from "../../../state/transactionsSlice";
-import Button from "../../../components/Button";
+import { useBackd } from "../../../../app/hooks/use-backd";
+import { AppDispatch } from "../../../../app/store";
+import Popup from "../../../../components/Popup";
+import { Position, Pool, TransactionInfo, fromPlainPosition } from "../../../../lib/types";
+import { removePosition } from "../../../../state/positionsSlice";
+import { selectTransactions } from "../../../../state/transactionsSlice";
+import Button from "../../../../components/Button";
 
 const ButtonContainer = styled.div`
   width: 100%;
