+++ src/pages/actions/register/topup/LoanSearch.tsx
@@ -4,16 +4,16 @@ import { useTranslation } from "react-i18next";
 import styled from "styled-components";
 import { useDispatch, useSelector } from "react-redux";
 
-import { useBackd } from "../../../app/hooks/use-backd";
-import BasicInput from "../../../components/BasicInput";
-import { spinAnimation } from "../../../styles/animations/SpinAnimation";
-import pending from "../../../assets/ui/status/pending.svg";
-import RowSelector from "../../../components/RowSelector";
-import { selectEthPrice } from "../../../state/poolsListSlice";
-import { Loan } from "../../../lib/types";
-import { fetchLoans, selectLoans } from "../../../state/lendingSlice";
-import { RowOptionType } from "../../../components/RowOption";
-import { AppDispatch } from "../../../app/store";
+import { useBackd } from "../../../../app/hooks/use-backd";
+import BasicInput from "../../../../components/BasicInput";
+import { spinAnimation } from "../../../../styles/animations/SpinAnimation";
+import pending from "../../../../assets/ui/status/pending.svg";
+import RowSelector from "../../../../components/RowSelector";
+import { selectEthPrice } from "../../../../state/poolsListSlice";
+import { Loan } from "../../../../lib/types";
+import { fetchLoans, selectLoans } from "../../../../state/lendingSlice";
+import { RowOptionType } from "../../../../components/RowOption";
+import { AppDispatch } from "../../../../app/store";
 
 const StyledLoanSearch = styled.div`
   width: 100%;
