+++ src/pages/actions/register/topup/ProtectableLoan.tsx
@@ -1,29 +1,24 @@
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 import styled from "styled-components";
-import { useNavigate, useParams } from "react-router-dom";
+import { useWeb3React } from "@web3-react/core";
+import { useNavigate } from "react-router-dom";
 
-import { selectLoans } from "../../../state/lendingSlice";
-import { selectEthPrice } from "../../../state/poolsListSlice";
-import Asset from "../../../components/Asset";
-import { selectPool } from "../../../state/selectors";
-import { GradientText } from "../../../styles/GradientText";
-import { LendingProtocol, Loan } from "../../../lib/types";
-import { TOPUP_ACTION_ROUTE } from "../../../lib/constants";
+import { selectEthPrice } from "../../../../state/poolsListSlice";
+import Button from "../../../../components/Button";
+import { Loan } from "../../../../lib/types";
+import { TOPUP_ACTION_ROUTE } from "../../../../lib/constants";
 
-const StyledActionSummary = styled.div`
+const StyledProtectableLoan = styled.div`
   width: 100%;
   display: flex;
   justify-content: space-between;
   align-items: center;
   background: rgba(255, 255, 255, 0.1);
+  border: 1px solid rgba(255, 255, 255, 0.2);
   border-radius: 1.4rem;
   padding: 1.3rem 1.4rem;
-
   margin-top: 1rem;
-  @media (max-width: 600px) {
-    margin-top: 0.6rem;
-  }
 `;
 
 interface ColumnProps {
@@ -36,6 +31,7 @@ const Column = styled.div`
   flex-direction: column;
 
   @media (max-width: 600px) {
+    flex: auto;
     display: ${(props: ColumnProps) => (props.hideMobile ? "none" : "flex")};
   }
 `;
@@ -60,12 +56,10 @@ interface ValueProps {
 const Value = styled.div`
   font-weight: 700;
   letter-spacing: 0.2px;
-  display: flex;
-  align-items: center;
 
   font-size: 1.8rem;
   @media (max-width: 600px) {
-    font-size: 1.4rem;
+    font-size: 1.6rem;
   }
 
   @media only percy {
@@ -73,35 +67,25 @@ const Value = styled.div`
   }
 `;
 
-const ChangePoolButton = styled.button`
-  margin-left: 0.5rem;
-  cursor: pointer;
-`;
+interface Props {
+  loan: Loan;
+}
 
-const ChangePoolText = styled(GradientText)`
-  font-weight: 500;
-  font-size: 1rem;
-`;
-
-const ActionSummary = (): JSX.Element => {
+const ProtectableLoan = ({ loan }: Props): JSX.Element => {
   const { t } = useTranslation();
-  const { address, protocol, poolName } = useParams<"address" | "protocol" | "poolName">();
+  const { account } = useWeb3React();
   const navigate = useNavigate();
   const ethPrice = useSelector(selectEthPrice);
-  const pool = useSelector(selectPool(poolName));
-  const loans = useSelector(selectLoans(address));
 
-  const loan = loans.filter((loan: Loan) => loan.protocol === protocol)[0];
-
-  if (!loan || !pool) return <div />;
+  if (!loan) return <div />;
 
   return (
-    <StyledActionSummary id="action-summary">
+    <StyledProtectableLoan id={`${loan.protocol.toLowerCase()}-protectable-loan`}>
       <Column>
         <Header>{t("actions.suggestions.topup.labels.protocol")}</Header>
         <Value>{loan.protocol}</Value>
       </Column>
-      <Column>
+      <Column hideMobile>
         <Header>{t("actions.suggestions.topup.labels.healthFactor")}</Header>
         <Value hideOnSnapshot>{loan.healthFactor.toCryptoString()}</Value>
       </Column>
@@ -114,19 +98,16 @@ const ActionSummary = (): JSX.Element => {
         <Value hideOnSnapshot>{loan.totalDebtETH.toUsdValue(ethPrice)}</Value>
       </Column>
       <Column>
-        <Header>{t("actions.suggestions.topup.labels.pool")}</Header>
-        <Value>
-          <Asset tiny token={pool.underlying} />
-          <ChangePoolButton
-            id="action-summary-change-pool"
-            onClick={() => navigate(`${TOPUP_ACTION_ROUTE}/${address}/${protocol}`)}
-          >
-            <ChangePoolText>{t("components.change")}</ChangePoolText>
-          </ChangePoolButton>
-        </Value>
+        <Button
+          id={`${loan.protocol.toLowerCase()}-protectable-loan-button`}
+          medium
+          text={t("actions.suggestions.topup.register")}
+          background="#3A3550"
+          click={() => navigate(`${TOPUP_ACTION_ROUTE}/${account}/${loan.protocol}`)}
+        />
       </Column>
-    </StyledActionSummary>
+    </StyledProtectableLoan>
   );
 };
 
-export default ActionSummary;
+export default ProtectableLoan;
