+++ src/pages/actions/register/topup/RegisterTopupConfirmation.tsx
@@ -2,12 +2,12 @@ import { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useTranslation } from "react-i18next";
 
-import { useBackd } from "../../../app/hooks/use-backd";
-import { AppDispatch } from "../../../app/store";
-import Popup from "../../../components/Popup";
-import { registerPosition } from "../../../state/positionsSlice";
-import { Pool, Position } from "../../../lib/types";
-import { hasPendingTransaction } from "../../../state/transactionsSlice";
+import { useBackd } from "../../../../app/hooks/use-backd";
+import { AppDispatch } from "../../../../app/store";
+import Popup from "../../../../components/Popup";
+import { registerPosition } from "../../../../state/positionsSlice";
+import { Pool, Position } from "../../../../lib/types";
+import { hasPendingTransaction } from "../../../../state/transactionsSlice";
 import TopupInformation from "./TopupInformation";
 
 interface Props {
